Separated MainPage from MainWindow to access resources.
